+++ src/text/ascii/digits/mod.rs
@@ -9,8 +9,6 @@ mod _test;
 mod _docs; // DOC_*
 use _docs::*;
 
-mod _helper; // ascii_digit_*
-
 mod define; // Digits
 mod u8;
 mod u16;
@@ -19,13 +17,10 @@ mod u64;
 mod u128;
 mod usize;
 
-crate::structural_mods! { // _mods, _crate_internals
+crate::structural_mods! { // _mods
     _mods {
         pub use super::{
             define::Digits,
         };
-    }
-    _crate_internals {
-        pub(crate) use super::_helper::*;
     }
 }