+++ src/text/mod.rs
@@ -46,7 +46,6 @@ crate::structural_mods! { // _mods, _pub_mods, _crate_internals
     _crate_internals {
         pub(crate) use super::{
             _DOC_TEXT_MODULES,
-            ascii::_crate_internals::*,
         };
     }
     _hidden {