+++ src/text/unicode/scalar/namespace/char.rs
@@ -3,7 +3,7 @@
 // TOC
 // - methods over char
 
-use crate::{AsciiLut, Char, Pcg32, is, unwrap};
+use crate::{Ascii, Char, Pcg32, is, unwrap};
 
 /// # Methods over `char`
 #[rustfmt::skip]
@@ -87,14 +87,14 @@ impl Char<char> {
     /// Returns the ASCII representation as a `&'static str`, or `""` if non-ASCII.
     #[must_use]
     pub const fn as_ascii(self) -> &'static str {
-        is![self.0.is_ascii(), AsciiLut::ASCII_CHARS[self.0 as usize], ""]
+        is![self.0.is_ascii(), Ascii::ASCII_CHARS[self.0 as usize], ""]
     }
 
     /// Returns the ASCII representation as a `&'static str`.
     /// # Panics
     /// Panics if the character is not ASCII.
     #[must_use]
-    pub const fn as_ascii_unchecked(self) -> &'static str { AsciiLut::ASCII_CHARS[self.0 as usize] }
+    pub const fn as_ascii_unchecked(self) -> &'static str { Ascii::ASCII_CHARS[self.0 as usize] }
 
     /// Converts a character to its closest ASCII equivalent, if possible.
     ///